% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.pb.umi}
\alias{get.pb.umi}
\title{Creates a pseudobulk using a grouped tibble}
\usage{
get.pb.umi(umi, df.grouped, cell.col = "cell_id", min.cells = 50,
  n.cores = 1)
}
\arguments{
\item{umi}{UMI count matrix (genes x cells)}

\item{df.grouped}{Grouped tibble. The grouping should only be based on a single column}

\item{cell.col}{Column in the tibble, where the cell names are stored.}

\item{min.cells}{Minimal number of cells which should be merged to a pseudobulk}

\item{n.cores}{Number of cores to run in parallel.}
}
\value{
Matrix of pseudobulks
}
\description{
Creates a pseudobulk using a grouped tibble
}
