% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge.sparse}
\alias{merge.sparse}
\title{Merge sparse matrices}
\usage{
\method{merge}{sparse}(...)
}
\arguments{
\item{...}{List of sparse matrices}
}
\value{
Merged sparse matrix
}
\description{
Merge sparse matrices
}
