% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.topGO}
\alias{run.topGO}
\title{GO enrichment of selected genes}
\usage{
run.topGO(interesting.genes, gene.universe, gene2go.list, n.out = 25)
}
\arguments{
\item{interesting.genes}{Genes to test enrichment for}

\item{gene.universe}{All genes to test against}

\item{gene2go.list}{List of gene-name to GO-term}

\item{n.out}{Number of rows to return of the tibble}
}
\value{
Tibble with enrichment analysis, based on Fisher exact test.
}
\description{
GO enrichment of selected genes
}
